remove signup section/nav, remove donator section
There is now a signup button first thing when you access the site, and I
removed it from the navbar because it was not needed anymore. I also
moved donators to the sponsors section because it was taking up too much
real estate. These will eventually go to their own respective pages